cauliflower_ai-main/cauliflower_ai-main/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==4.2.23
django-browser-reload==1.18.0
filelock==3.19.1
fsspec==2025.9.0
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
packaging==25.0
pillow==11.3.0
PyYAML==6.0.2
requests==2.32.5
safetensors==0.6.2
sqlparse==0.5.3
sympy==1.14.0
timm==1.0.19
torch==2.8.0+cpu
torchvision==0.23.0+cpu
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
